Pass itself and a user-defined value to processor
This allows storing of metadata, which allows the processor to access
other data related to its own value. You can use this to refer to other
members of the parent object, which will be needed in tvec. However, it
is not needed right now for normal vec2.

The proxy object itself is also passed to allow usage of the other
processor and hacking to get the underlying value by casting it to T,
since it is the first member of this class.