Create bind_first in utilities
Create utils.hpp for utilities. First one is a function to bind the
first argument of a std::function. Can be used to bind a member
function, which is being used in the ongoing sigslots implementation.

Due to bind_first needing std::integer_sequence, changed the c++ version in
the Tupfile to c++14 from c++11.

Thank you to T.C. on stackoverflow for the bind_first implementation!
From http://stackoverflow.com/a/26130700.