Create keyboard input interface
The keystate class gets the true current state of the key, while the
keyinput class is bound to the update() method, and detects if the state
has changed between refreshes.

Note that keyinput might still miss keypressed if the key was pressed
and released between presses. Therefore, the window may have to handle
these.